Pageable

+++ src/main/java/br/senai/sc/engajamento/comentario/service/ComentarioService.java
@@ -7,13 +7,15 @@ import br.senai.sc.engajamento.exception.AcaoNaoPermitidaException;
 import br.senai.sc.engajamento.exception.NaoEncontradoException;
 import br.senai.sc.engajamento.usuario.model.entity.Usuario;
 import br.senai.sc.engajamento.usuario.repository.UsuarioRepository;
-import br.senai.sc.engajamento.usuario.service.UsuarioService;
 import br.senai.sc.engajamento.video.model.entity.Video;
 import br.senai.sc.engajamento.video.repository.VideoRepository;
 import br.senai.sc.engajamento.video.service.VideoService;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
-import lombok.Data;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.*;
@@ -47,43 +49,63 @@ public class ComentarioService {
     }
 
     public Comentario buscarUm(@Valid BuscarUmComentarioCommand cmd) {
-        return retornaComentario(cmd.getIdComentario());
+        Comentario comentario = retornaComentario(cmd.getIdComentario());
+        if(!comentario.getIdVideo().getEhInativado()){
+            return retornaComentario(cmd.getIdComentario());
+        }
+        throw new NaoEncontradoException("Vídeo não encontrado");
     }
 
-    public List<Comentario> buscarTodosPorVideo(@Valid BuscarTodosPorVideoComentarioCommand cmd) {
+    public Page<Comentario> buscarTodosPorVideo(
+            @Valid BuscarTodosPorVideoComentarioCommand cmd, int page) {
         Video video = videoRepository.getById(cmd.getIdVideo());
-        List<Comentario> list = comentarioRepository.findAllByIdVideo(video);
-        list.sort(Comparator.comparing(Comentario::getDataHora).reversed());
-        return list;
+        if(!video.getEhInativado()){
+            Pageable pageable = PageRequest.of(page, 5);
+            Page<Comentario> list = comentarioRepository.findAllByIdVideoOrderByDataHora(video, pageable);
+            return list;
+        }
+        throw new NaoEncontradoException("Vídeo não encontrado");
     }
 
-    public List<Comentario> buscarTodosPorData(@Valid BuscarTodosPorDataComentarioCommand cmd)  {
+    public Page<Comentario> buscarTodosPorData(@Valid BuscarTodosPorDataComentarioCommand cmd, int page)  {
         List<Comentario> listaComentariosFiltrados = new ArrayList<>();
         Video video = videoRepository.getById(cmd.getIdVideo());
 
-        List<Comentario> listaComentarios =
-                comentarioRepository.findAllByIdVideo(video);
-        for (Comentario comentario : listaComentarios) {
-            if (comentario.getDataHora().toLocalDate().equals(cmd.getData())) {
-                listaComentariosFiltrados.add(comentario);
+        if(!video.getEhInativado()){
+            Pageable pageable = PageRequest.of(page, 5);
+            List<Comentario> listaComentarios =
+                    comentarioRepository.findAllByIdVideo(video);
+            for (Comentario comentario : listaComentarios) {
+                if (comentario.getDataHora().toLocalDate().equals(cmd.getData())) {
+                    listaComentariosFiltrados.add(comentario);
+                }
             }
+
+            return converteListParaPage(listaComentariosFiltrados, pageable);
         }
-        return listaComentariosFiltrados;
+        throw new NaoEncontradoException("Vídeo não encontrado");
     }
 
     public Integer buscarQuantidadeRespostas(@Valid BuscarQuantidadeRepostasComentarioCommand cmd) {
-        return retornaComentario(cmd.getIdComentario()).getQtdRespostas();
+        Comentario comentario = retornaComentario(cmd.getIdComentario());
+        if(!comentario.getIdVideo().getEhInativado()){
+            return retornaComentario(cmd.getIdComentario()).getQtdRespostas();
+        }
+        throw new NaoEncontradoException("Vídeo não encontrado");
     }
 
     public void deletar(@Valid DeletarComentarioCommand cmd) {
         Comentario comentario = retornaComentario(cmd.getIdComentario());
         Video video = videoRepository.getById(comentario.getIdVideo().getId());
-        if (!(cmd.getIdUsuario().equals(comentario.getIdUsuario().getIdUsuario()))) {
-            throw new AcaoNaoPermitidaException();
+        if(!video.getEhInativado()){
+            if (!(cmd.getIdUsuario().equals(comentario.getIdUsuario().getIdUsuario()))) {
+                throw new AcaoNaoPermitidaException();
+            }
+            video.setQtdComentarios(video.getQtdComentarios() - 1);
+            videoRepository.save(video);
+            comentarioRepository.delete(comentario);
         }
-        video.setQtdComentarios(video.getQtdComentarios() - 1);
-        videoRepository.save(video);
-        comentarioRepository.delete(comentario);
+        throw new NaoEncontradoException("Vídeo não encontrado");
     }
 
     public Comentario retornaComentario(String idComentario) {
@@ -92,6 +114,18 @@ public class ComentarioService {
             return optional.get();
         }
         throw new NaoEncontradoException("Comentário não encontrado!");
+    }
+
+    public Page<Comentario> converteListParaPage(List<Comentario> list, Pageable pageable) {
+        // Obtém o índice de início da página com base no offset do Pageable
+        int start = (int) pageable.getOffset();
+
+        // Calcula o índice final da página, levando em consideração o tamanho da página
+        int end = Math.min((start + pageable.getPageSize()), list.size());
+
+        // Cria uma nova instância de PageImpl com a parte relevante da lista,
+        // as informações de paginação e o tamanho total da lista original
+        return new PageImpl<>(list.subList(start, end), pageable, list.size());
     }
 
 }
